+++ packages/superdoc/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.9.0-next.9](https://github.com/superdoc-dev/superdoc/compare/v1.9.0-next.8...v1.9.0-next.9) (2026-01-28)
+
+
+### Bug Fixes
+
+* patch broken numbering definitions ([#1848](https://github.com/superdoc-dev/superdoc/issues/1848)) ([f34b121](https://github.com/superdoc-dev/superdoc/commit/f34b1217eb7afc01f49da8db54d63e6e3126185a))
+
+# [1.9.0-next.8](https://github.com/superdoc-dev/superdoc/compare/v1.9.0-next.7...v1.9.0-next.8) (2026-01-28)
+
+
+### Bug Fixes
+
+* annotation and interaction issues ([#1847](https://github.com/superdoc-dev/superdoc/issues/1847)) ([ffb1055](https://github.com/superdoc-dev/superdoc/commit/ffb1055dcf21916a23b221655a9eff8828d49fa0))
+
+# [1.9.0-next.7](https://github.com/superdoc-dev/superdoc/compare/v1.9.0-next.6...v1.9.0-next.7) (2026-01-27)
+
+
+### Bug Fixes
+
+* **export:** prefix Relationship IDs with rId for valid xsd:ID ([#1855](https://github.com/superdoc-dev/superdoc/issues/1855)) ([11e67e1](https://github.com/superdoc-dev/superdoc/commit/11e67e1e4332976df279edf62f1aa33177004f9e))
+
 # [1.9.0-next.5](https://github.com/superdoc-dev/superdoc/compare/v1.9.0-next.4...v1.9.0-next.5) (2026-01-24)